Multiple strava accounts - passing tests and cleanup

+++ application/strava_api/views.py
@@ -8,7 +8,6 @@ from flask_login import current_user, login_required
 import pandas as pd
 from scipy.interpolate import interp1d
 from sqlalchemy.exc import IntegrityError
-from stravalib import Client
 
 from . import strava_api
 from application import converters, util
@@ -31,7 +30,7 @@ def authorize():
     'http://localhost:5000'
   )
 
-  return redirect(Client().authorization_url(
+  return redirect(StravaAccount.get_client().authorization_url(
     CLIENT_ID,
     scope=['activity:read_all'],
     redirect_uri=urljoin(
@@ -74,13 +73,13 @@ def handle_code():
       error=error
     )
 
-  token = Client().exchange_code_for_token(
+  token = StravaAccount.get_client().exchange_code_for_token(
     client_id=CLIENT_ID,
     client_secret=CLIENT_SECRET,
     code=request.args.get('code'),
   )
 
-  athlete = Client(access_token=token['access_token']).get_athlete()
+  athlete = StravaAccount.get_client(access_token=token['access_token']).get_athlete()
 
   strava_acct = StravaAccount(
     strava_id=athlete.id,
@@ -105,15 +104,13 @@ def handle_code():
 @login_required
 def display_activity_list():
   """Display list of strava activities to view in individual Dashboards."""
-  try:
-    strava_account = StravaAccount.query.get(request.args.get('id'))
-  except IntegrityError as e:
-    print(e)
+  strava_account = StravaAccount.query.get(request.args.get('id'))
+
+  if strava_account is None:
     return redirect(url_for('strava_api.authorize'))
     # '?after="/strava/activities"'
 
-  token = strava_account.get_token()
-  client = Client(access_token=token['access_token']) # strava_account.token[]
+  client = strava_account.client
 
   limit = int(request.args.get('limit', 25))
   page = int(request.args.get('page', 1))
@@ -122,11 +119,11 @@ def display_activity_list():
   activities.per_page = limit
   activities._page = page
   activity_list = list(activities)
+  saved_activity_id_list = [a.strava_id for a in strava_account.activities.all()]
 
   for activity in activity_list:
-    print([a.strava_id == activity.id for a in strava_account.activities])
-    activity._saved = any([a.strava_id == activity.id for a in strava_account.activities])
-    # if Activity.query.filter_by(activity.strava_id)
+    # activity._saved = any([a.strava_id == activity.id for a in strava_account.activities])
+    activity._saved = activity.id in saved_activity_id_list
 
   form = BatchForm()
   if form.validate_on_submit():
@@ -218,15 +215,20 @@ def manage():
 @strava_api.route('/revoke')
 @login_required
 def revoke():
-  try:
-    strava_account = StravaAccount.query.get(request.args.get('id'))
-  except IntegrityError as e:
-    print(e)
+  
+  strava_account = StravaAccount.query.get(request.args.get('id'))
+  
+  if strava_account is None:
+    flash(f'No strava account was found with id {request.args.get("id")}')
+    return redirect(url_for('strava_api.manage'))
+
+  msg_success = (
+    f'Strava account {strava_account.firstname} {strava_account.lastname} '
+     'successfully removed!'
+  )
 
   db.session.delete(strava_account)
   db.session.commit()
 
-  flash(f'Strava account {strava_account.firstname} {strava_account.lastname} '
-         'successfully removed!')
-
+  flash(msg_success)
   return redirect(url_for('strava_api.manage'))
